Removed connection reference counting and simplified destructors. Also added some error handling to destructors.
Reference counting is not needed anymore as libvirt_get_connect does not create new resource handle anymore. This was the only place where we could get more resource handles for one connection.
PHP is calling destructors in correct (reverse in meaning of PHP script) order. This solves the problem when we did not close libvirt connection properly because of active domain resources.
We need to be sure that we closed all domains with virDomainFree before calling virConnectClose or it will fail and we will not close the connection and we will eat up all available connections to libvirt :-)